+++ NEWS.md
@@ -3,6 +3,7 @@
 - new function label_outliers to extend boxplots
 - new function detect_outliers for compatibility with geom_boxplot
 - new function WINratio to calculate the ratio of wins to losses in clinical trials
+- new function flex2rmd to convert flextable to rmarkdown table if non-interactive
 
 # wrappedtools 0.9.3
 - compare2numvars checks indep. variables for 2 levels